require auth for all pages expect auth
